# swagger.yaml
openapi: 3.0.0
info:
  title: GitHub Documentation Extraction API
  description: API for extracting and processing documentation from GitHub repositories
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: /
    description: Default server
paths:
...
  /api/v1/repos/{owner}/{repo}:
    get:
      summary: Get repository information
      description: Fetch information about a GitHub repository and optionally its documentation
      operationId: getRepository
      tags:
        - Repositories
      parameters:
        - name: owner
          in: path
          required: true
          description: Repository owner/organization
          schema:
            type: string
        - name: repo
          in: path
          required: true
          description: Repository name
          schema:
            type: string
        - name: skip_docs
          in: query
          required: false
          description: Skip fetching documentation
          schema:
            type: boolean
            default: false
        - name: force_refresh
          in: query
          required: false
          description: Force refresh the repository data
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Repository information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Repository:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: GitHub repository ID
        name:
          type: string
          description: Repository name
        full_name:
          type: string
          description: Full repository name (owner/repo)
        description:
          type: string
          description: Repository description
        stars:
          type: integer
          description: Number of stars
        forks:
          type: integer
          description: Number of forks
        language:
          type: string
          description: Primary programming language
        topics:
          type: array
          items:
            type: string
          description: Repository topics/tags
        default_branch:
          type: string
          description: Default branch name
        created_at:
          type: string
          format: date-time
          description: Repository creation date
        updated_at:
          type: string
          format: date-time
          description: Repository last update date
        url:
          type: string
          description: API URL for the repository
        html_url:
          type: string
          description: Web URL for the repository
        documentation_url:
          type: string
          description: URL for documentation if available
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        status:
          type: integer
          description: HTTP status code
...